requirements.txt
# Core Python dependencies for the Gen–Ver DAgger pipeline.
# CUDA toolkits, _libgcc_mutex, and other Conda-only packages must be
# installed via the README instructions (conda env + NVIDIA drivers).

accelerate==1.11.0
datasets==4.4.1
evaluate==0.4.6
hydra-core==1.3.2
omegaconf==2.3.0
openai>=1.61.0
pandas==2.3.3
peft==0.17.1
pyarrow==22.0.0
rich==14.2.0
azure-identity>=1.19.0
flash-attn>=2.7.4
safetensors==0.6.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
transformers==4.57.1
trl @ git+https://github.com/huggingface/trl.git@69ad852e5654a77f1695eb4c608906fe0c7e8624
tqdm==4.67.1
vllm==0.11.0
wandb==0.22.3
